moved Pwd to pgp-password
