+++ extension/js/common/core/pgp-password.ts
@@ -5,6 +5,8 @@
 import { openpgp } from './pgp.js';
 import { base64encode, secureRandomBytes } from '../platform/util.js';
 
+export type Pwd = { question?: string; answer: string; };
+
 interface PasswordStrengthResult {
   word: {
     match: string;
